Added tests for tokenizer
